+++ NEWS.md
@@ -5,6 +5,12 @@
 
 - `osmdata_data_frame` adds columns `osm_center_lat` and `osm_center_lon` for `out * center;` queries (#316).
 
+## Minor changes
+
+- Fix queries with `!match_case` and only one value (#317)
+- Fix queries with multiple features & multiple osm_types (#318)
+
+
 0.2.1
 ===================